backend/requirements.txt
Flask==3.0.3
requests==2.32.3
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
pymongo==4.10.1
gunicorn==21.2.0
APScheduler==3.10.4
PyPDF2==3.0.1
google-cloud-storage==2.19.0
